+++ changelog.md
@@ -1,4 +1,7 @@
 
+# 14-11-2025 :
+- added `get_child_index` method to `element` class in `Dom.js` file
+
 # 12-11-2025 :
 - added `content_editable, get_content_editable` methods to `element` class in `Dom.js` file